+++ src/client/mod.rs
@@ -4,6 +4,7 @@ pub use self::{
     executor_ext::ExecutorExt,
     stream::Stream,
     transaction::{Transaction, TransactionBuilder},
+    tuple_response::TupleResponse,
 };
 
 pub mod schema;
@@ -13,6 +14,7 @@ mod executor;
 mod executor_ext;
 mod stream;
 mod transaction;
+mod tuple_response;
 
 mod private {
     use crate::client::{Connection, Stream, Transaction};